Functionality: Get all records + take attendance

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/repository/ScannerRepository.java
@@ -2,7 +2,9 @@ package edu.miu.cs.cs544.repository;
 
 import edu.miu.common.repository.BaseRepository;
 import edu.miu.cs.cs544.domain.Scanner;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestParam;
 
-public interface ScannerRepository extends BaseRepository<Scanner, Integer>{
-
+public interface ScannerRepository extends BaseRepository<Scanner, Long>{
 }
